\name{gergnpsa}
 \docType{data}
  \alias{gergnpsa}
  \title{Real GNP in Germany (1960.1-1990.4). Seasonally Adjusted}

  \description{
    Real GNP in Germany. Sample: 1960.1 - 1990.4. Remark: Seasonally Adjusted. This time series is part of
    the data set used in the book Franses (1996).
  }

  \format{An object of class 'ts' containing the data.}


  \usage{
    data ("gergnpsa")
  }

  \source{P.H. Franses's homepage: \url{http://www.few.eur.nl/few/people/franses/}.}

  \references{
    P.H. Franses: Periodicity and Stochastic Trends in Economic Time Series (Oxford University Press, 1996).

    J. Wolters (1992): Persistence and Seasonality in Output and Employment of the Federal Republic of
    Germany, \emph{Recherches Economiques de Louvain}, \bold{58} pp.421-39.    
  }

\keyword{datagen}


