\name{swdipc}
 \docType{data}
  \alias{swdipc}
  \title{Real per Capita Disposable Income in Sweden (1963.1-1988.1)}

  \description{
    Real per capita disposable income in Sweden. Sample: 1963.1 - 1988.1. Remark: measured in logs. This time     series is part of the data set used in the book Franses (1996).
  }

  \format{An object of class 'ts' containing the data.}      

  \usage{
    data ("swdipc")
  }

  \source{Previously available on P.H. Franses's homepage: \url{https://www.eur.nl/en/people/philip-hans-franses}.}

  \references{
    B. Assarson (1991): 'The Stochastic Behaviour of Durables and Nondurables Consumption in Sweden', Working 
    paper 1991:21, Department of Economics, Uppsala University.

    P.H. Franses: Periodicity and Stochastic Trends in Economic Time Series (Oxford University Press, 1996).
  }

\keyword{datagen}


