\name{ukinvest}
 \docType{data}
  \alias{ukinvest}
  \title{Real Total Investment in the United Kindom (1955.1-1988.4)}

  \description{
    Real total investment in the United Kindom. Sample: 1955.1 - 1988.4. This time series is part of the data
    set used in the book Franses (1996). 
  }

  \format{An object of class 'ts' containing the data.}

  \usage{
    data ("ukinvest")
  }

  \source{Previously available on P.H. Franses's homepage: \url{https://www.eur.nl/en/people/philip-hans-franses}.}

  \references{
    P.H. Franses: Periodicity and Stochastic Trends in Economic Time Series (Oxford University Press, 1996).

    D.R. Osborn (1990): A Survey of Seasonality in UK Macroeconomic Variables, \emph{International Journal of
    Forecasting}, \bold{6}, pp. 327-36.
  }

\keyword{datagen}


