\name{ukndcons}
 \docType{data}
  \alias{ukndcons}
  \title{United Kindom non-durables Consumption (1955.1-1988.4)}

  \description{
    United Kindom nondurables consumption. Sample: 1955.1 - 1988.4. Remark: at 1985 prices. This time series
    is part of the data set used in the book Franses (1996).
  }

  \format{An object of class 'ts' containing the data.}

  \usage{
    data ("ukndcons")
  }

  \source{P.H. Franses's homepage: \url{http://www.few.eur.nl/few/people/franses/}.}

  \references{
    P.H. Franses: Periodicity and Stochastic Trends in Economic Time Series (Oxford University Press, 1996).

    D.R. Osborn (1990): A Survey of Seasonality in UK Macroeconomic Variables, \emph{International Journal of
    Forecasting}, \bold{6}, pp. 327-36.
  }

\keyword{datagen}


