\name{usaipi}
 \docType{data}
  \alias{usaipi}
  \title{Total Industrial Production Index for the United States (1960.1-1991.4)}

  \description{
    OECD Main Economic Indicators: Total industrial production index for the United States. Sample: 1960.1 -
    1991.4. Remark: (1985=100). This time series is part of the data set used in the book Franses (1996).
  }

  \format{An object of class 'ts' containing the data.}  

  \usage{
    data ("usaipi")
  }

  \source{P.H. Franses's homepage: \url{http://www.few.eur.nl/few/people/franses/}.}

  \references{
    P.H. Franses: Periodicity and Stochastic Trends in Economic Time Series (Oxford University Press, 1996).
  }

\keyword{datagen}


