\name{canunsa}
 \docType{data}
  \alias{canunsa}
  \title{Unemployment in Canada. (1960.1-1987.4). Seasonally Adjusted}

  \description{
    OECD Main Economic Indicators: Unemployment in Canada. Sample: 1960.1 - 1987.4. Remark: Seasonally
    Adjusted. This time series is part of the data set used in the book Franses (1996).
  }

  \format{An object of class 'ts' containing the data.}

  \usage{
    data ("canunsa")
  }

  \source{P.H. Franses's homepage: \url{http://www.few.eur.nl/few/people/franses/}.}

  \references{
    P.H. Franses: Periodicity and Stochastic Trends in Economic Time canunsas (Oxford University Press, 1996).
  }

\keyword{datagen}


