\name{usaipisa}
 \docType{data}
  \alias{usaipisa}
  \title{Total Industrial Production Index for the United States (1960.1-1991.4). Seasonally Adjusted}

  \description{
    OECD Main Economic Indicators: Total industrial production index for the United States. Sample: 1960.1 -
    1991.4. Remark: (1985=100), Seasonally Adjusted. This time series is part of the data set used in the
    book Franses (1996).
  }

  \format{An object of class 'ts' containing the data.}  

  \usage{
    data ("usaipisa")
  }

  \source{Previously available on P.H. Franses's homepage: \url{https://www.eur.nl/en/people/philip-hans-franses}.}

  \references{
    P.H. Franses: Periodicity and Stochastic Trends in Economic Time Series (Oxford University Press, 1996).
  }

\keyword{datagen}


